+++ packages/cli/src/commands/server.ts
@@ -4,12 +4,10 @@ import { defaultsEnv } from '../helper'
 
 export function action(opts: {
   port?: number
-  cache?: boolean
 }): void {
   defaultsEnv()
 
   const server = new Server(process.env.FaasRoot, {
-    cache: opts.cache,
     port: opts.port || 3000,
   })
 
@@ -20,13 +18,12 @@ export function ServerCommand(program: Command): void {
   program
     .command('server')
     .name('server')
-    .description('Start local server')
+    .description('Start faas server')
     .on('--help', () => {
       console.log(`
 Examples:
   npm exec faas server`)
     })
     .option('-p, --port <port>', 'Port', '3000')
-    .option('-c, --cache', 'Cache functions', false)
     .action(action)
 }