+++ packages/server/src/index.ts
@@ -11,6 +11,24 @@
  * npm install @faasjs/server tsx
  * ```
  *
+ * ## Usage
+ *
+ * 1. Create a `server.ts` file:
+ * ```ts
+ * // server.ts
+ * import { Server } from '@faasjs/server'
+ *
+ * const server = new Server({
+ *  // options
+ * })
+ *
+ * server.start()
+ * ```
+ * 2. Run the server:
+ * ```sh
+ * tsx server.ts // or `bun server.ts` if you are not using bun
+ * ```
+ *
  * ## Routing
  *
  * Static routing:
@@ -25,14 +43,6 @@
  *
  * @packageDocumentation
  */
-
-import(
-  (globalThis as any).Bun
-    ? 'bun'
-    : typeof process.env.JEST_WORKER_ID === 'string'
-      ? 'jest'
-      : 'tsx'
-)
 
 export * from './middleware'
 export * from './server'