+++ packages/test/src/__tests__/basic.test.ts
@@ -1,8 +1,9 @@
 import { expect, it } from 'vitest'
 import { FuncWarper } from '../../src/index'
+import Func from './funcs/basic.func'
 
 it('basic', async () => {
-  const func = new FuncWarper(await import('./funcs/basic.func'))
+  const func = new FuncWarper(Func)
   const res = await func.handler<boolean>({}, {})
 
   expect(res).toEqual(true)